+++ actors/sunbird-lms-mw/actors/user/src/main/java/org/sunbird/user/dao/UserSelfDeclarationDao.java
@@ -1,0 +1,21 @@
+package org.sunbird.user.dao;
+
+import java.util.List;
+import java.util.Map;
+import org.sunbird.common.request.RequestContext;
+import org.sunbird.models.user.UserDeclareEntity;
+
+public interface UserSelfDeclarationDao {
+  void insertSelfDeclaredFields(Map<String, Object> extIdMap, RequestContext context);
+
+  UserDeclareEntity upsertUserSelfDeclaredFields(
+      UserDeclareEntity userDeclareEntity, RequestContext context);
+
+  List<Map<String, Object>> getUserSelfDeclaredFields(
+      UserDeclareEntity userDeclareEntity, RequestContext context);
+
+  List<Map<String, Object>> getUserSelfDeclaredFields(String userId, RequestContext context);
+
+  void deleteUserSelfDeclaredDetails(
+      String userId, String orgId, String persona, RequestContext context);
+}
