+++ actors/sunbird-lms-mw/actors/user/src/main/java/org/sunbird/user/util/UserUtil.java
@@ -726,7 +726,7 @@ public class UserUtil {
           new UserDeclareEntity(
               (String) requestMap.get(JsonKey.USER_ID),
               prevOrgId,
-              JsonKey.TEACHER_PERSONA,
+              JsonKey.DEFAULT_PERSONA,
               userInfo);
       userDeclareEntity.setUpdatedBy((String) requestMap.get(JsonKey.UPDATED_BY));
       userDeclareEntity.setOperation(JsonKey.REMOVE);
@@ -754,7 +754,7 @@ public class UserUtil {
           new UserDeclareEntity(
               (String) requestMap.get(JsonKey.USER_ID),
               currOrgId,
-              JsonKey.TEACHER_PERSONA,
+              JsonKey.DEFAULT_PERSONA,
               userInfo);
       userDeclareEntity.setCreatedBy((String) requestMap.get(JsonKey.CREATED_BY));
       userDeclareEntity.setUpdatedBy((String) requestMap.get(JsonKey.UPDATED_BY));
